+++ src/samples/randoms/randoms_api/randoms_init_response.rs
@@ -3,11 +3,11 @@ use crate::id_targeted::IdTargeted;
 /// The response from a request to create a new Randoms struct within the thread pool
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct RandomsInitResponse {
-    pub id: u64,
+    pub id: usize,
 }
 
 impl IdTargeted for RandomsInitResponse {
-    fn id(&self) -> u64 {
+    fn id(&self) -> usize {
         self.id
     }
 }
